rename file to be consistent with other apps
