+++ apps/admin/frontend/src/app_root.test.tsx
@@ -6,7 +6,7 @@ import { LogSource, Logger } from '@votingworks/logging';
 import { screen } from '../test/react_testing_library';
 import { renderRootElement } from '../test/render_in_app_context';
 import { AppRoot } from './app_root';
-import { ApiMock, createApiMock } from '../test/helpers/api_mock';
+import { ApiMock, createApiMock } from '../test/helpers/mock_api_client';
 
 let apiMock: ApiMock;
