+++ apps/admin/frontend/src/components/reporting/filter_editor.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event';
 import { renderInAppContext } from '../../../test/render_in_app_context';
 import { FilterEditor } from './filter_editor';
 import { screen, within } from '../../../test/react_testing_library';
-import { ApiMock, createApiMock } from '../../../test/helpers/api_mock';
+import { ApiMock, createApiMock } from '../../../test/helpers/mock_api_client';
 
 let apiMock: ApiMock;
