+++ apps/admin/frontend/src/components/save_backend_file_modal.test.tsx
@@ -10,7 +10,7 @@ import { screen, waitFor } from '../../test/react_testing_library';
 import { SaveBackendFileModal } from './save_backend_file_modal';
 import { renderInAppContext } from '../../test/render_in_app_context';
 import { mockUsbDriveStatus } from '../../test/helpers/mock_usb_drive';
-import { ApiMock, createApiMock } from '../../test/helpers/api_mock';
+import { ApiMock, createApiMock } from '../../test/helpers/mock_api_client';
 
 let mockKiosk = fakeKiosk();
 let apiMock: ApiMock;
