+++ apps/admin/frontend/src/components/smartcard_modal/smartcard_modal.test.tsx
@@ -19,7 +19,7 @@ import {
   UserWithCard,
 } from '@votingworks/types';
 
-import { ApiMock, createApiMock } from '../../../test/helpers/api_mock';
+import { ApiMock, createApiMock } from '../../../test/helpers/mock_api_client';
 import { buildApp } from '../../../test/helpers/build_app';
 import { screen, waitFor, within } from '../../../test/react_testing_library';
 import { VxFiles } from '../../lib/converters';
