+++ apps/admin/frontend/src/screens/manual_data_entry_screen.test.tsx
@@ -8,7 +8,7 @@ import { hasTextAcrossElements } from '@votingworks/test-utils';
 import { screen, within } from '../../test/react_testing_library';
 import { renderInAppContext } from '../../test/render_in_app_context';
 import { ManualDataEntryScreen } from './manual_data_entry_screen';
-import { ApiMock, createApiMock } from '../../test/helpers/api_mock';
+import { ApiMock, createApiMock } from '../../test/helpers/mock_api_client';
 
 let apiMock: ApiMock;
