+++ apps/admin/frontend/src/screens/reports_screen.test.tsx
@@ -12,7 +12,7 @@ import { LogEventId, fakeLogger } from '@votingworks/logging';
 import { typedAs } from '@votingworks/basics';
 import { ReportsScreen } from './reports_screen';
 import { renderInAppContext } from '../../test/render_in_app_context';
-import { ApiMock, createApiMock } from '../../test/helpers/api_mock';
+import { ApiMock, createApiMock } from '../../test/helpers/mock_api_client';
 import { screen } from '../../test/react_testing_library';
 import { VxFiles } from '../lib/converters';
 import {
