+++ apps/admin/frontend/src/screens/unconfigured_screen.test.tsx
@@ -9,7 +9,7 @@ import { DEFAULT_SYSTEM_SETTINGS } from '@votingworks/types';
 import { renderInAppContext } from '../../test/render_in_app_context';
 import { screen } from '../../test/react_testing_library';
 
-import { ApiMock, createApiMock } from '../../test/helpers/api_mock';
+import { ApiMock, createApiMock } from '../../test/helpers/mock_api_client';
 import { UnconfiguredScreen } from './unconfigured_screen';
 
 let apiMock: ApiMock;
