+++ apps/admin/frontend/src/screens/write_in_adjudication_report_screen.test.tsx
@@ -9,7 +9,7 @@ import { fakeLogger, Logger } from '@votingworks/logging';
 import userEvent from '@testing-library/user-event';
 import { createMemoryHistory } from 'history';
 import { renderInAppContext } from '../../test/render_in_app_context';
-import { ApiMock, createApiMock } from '../../test/helpers/api_mock';
+import { ApiMock, createApiMock } from '../../test/helpers/mock_api_client';
 import { TallyWriteInReportScreen } from './write_in_adjudication_report_screen';
 import { screen, within } from '../../test/react_testing_library';
 import { routerPaths } from '../router_paths';
